pyproject.toml
[project]
name = "produtos-favoritos-api"
version = "0.1.0"
description = "API para gerenciamento de produtos favoritos de clientes"
authors = [
    {name = "Julia Dutra",email = "juliadutragoncalves@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "sqlalchemy[asyncio] (>=2.0.43,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-asyncio = "^1.1.0"
pytest-mock = "^3.15.0"
aiosqlite = "^0.21.0"
ruff = "^0.12.12"

[tool.ruff]
line-length = 120
exclude = [
    ".git",
    ".pytest_cache",
    ".venv",
    "__pycache__",
    "migrations",
    "app/api/v1/openapi_docs.py",
]
target-version = "py313"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP"]
ignore = ["B008"]

[tool.ruff.format]
quote-style = "double"